feat(port): reconcile v3 science onto release/0.8.0 + green check.sh (Phase 2)
Bring the diverged v3 science to feature-branch state (feature is the green v3 superset), reconciled against canonical's release machinery:
- Took feature's versions of harmonization/, values/, matching/, ingestion/, llm/, embedding/service, export/{eitl,visualization}, models/data_dictionary, clustering/subcluster, benchmarks/ (+aireadi) and their v3 unit tests.
- PRESERVED canonical machinery: __init__.py (importlib.metadata version-derivation), cli.py, pyproject.toml (0.7.0), test_cli/test_verify_release/test_smoke_test, publish.yml.
- cli.py v3 compat: re-added max_clusters to harmonize_leanb (the CLI cost cap, dropped in v3); widened _columns_for return so the **kwargs unpack matches v3 load_dictionary's detect_hierarchy: bool; type-ignore the click pass_context call.
- Excluded (GUI/integration, coordinate separately): test_ui_backend + twinsuk/matching/embedding/llm_batch integration tests.

check.sh green in the worktree: pytest 631 passed / 3 skipped, ruff + black clean, pyright src/ 0 errors (@ canonical pin 1.1.410).
Deferred follow-up: cohort data/examples + build_* scripts + the v2 notebook; version bump + CHANGELOG + build/twine + PR (user-gated at publish).

+++ src/ddharmon/export/visualization.py
@@ -237,7 +237,7 @@ def compute_umap_coords(
         random_state=random_state,
         n_neighbors=min(n_neighbors, len(vectors) - 1),
     )
-    return np.asarray(reducer.fit_transform(vectors))
+    return np.asarray(reducer.fit_transform(vectors), dtype=np.float64)
 
 
 def _shared_umap_layout(
